+++ polls/models.py
@@ -9,6 +9,7 @@ from static_content.utils import upload_to
 class Poll(models.Model):
     creator = models.ForeignKey(User, on_delete=models.CASCADE, related_name='polls')
     slug = models.SlugField(unique=True, max_length=255)
+    description = models.TextField(null=True, blank=True)
     title = models.CharField(max_length=255, unique=True)
     image = models.ImageField(upload_to=upload_to, null=True, blank=True)
     places_number = models.PositiveIntegerField()